+++ Psiphon-Browser/.gitignore
@@ -1,0 +1,1 @@
+psiphon_config.json
