+++ Psiphon-Browser/Browser/BrowserViewController.m
@@ -5,13 +5,13 @@
 //
 //
 
+#import <PsiphonTunnelController/Reachability.h>
 #import "BrowserViewController.h"
 #import "Tab.h"
 #import "BookmarksController.h"
 #import "BookmarksFormController.h"
 #import "BookmarkFolderFormController.h"
 #import "UIMainView.h"
-#import "Reachability.h"
 #import "Logging.h"
 
 #import <CoreFoundation/CoreFoundation.h>
